move CitationItem and its children to /common
