+++ src/common/components/CitationItem.jsx
