+++ src/common/components/PublicationInfoList.jsx
