+++ src/common/components/__tests__/PublicationInfo.test.jsx
