+++ src/common/components/__tests__/PublicationInfoList.test.jsx
